+++ CHANGES.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.1.0 - 2026-06-25
+### Added
+- `asimov-apple-calendar-emitter`
+- `asimov-apple-contacts-emitter`
+
 ## 0.0.1 - 2025-12-04
 ### Added
 - `asimov-apple-notes-emitter`